+++ android/app/src/main/java/com/globalradio/app/MainActivity.java
@@ -20,6 +20,7 @@ public class MainActivity extends BridgeActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
+        registerPlugin(BackgroundAudioPlugin.class);
         super.onCreate(savedInstanceState);
         setupToolbar();
     }